52 lesson ends here splitting widgets
