+++ lib/result.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class Result extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+        child: Text("You did it!")
+    );
+  }
+}
